move test runner to a subfolder
